[#3] fix : 화면 안뜨는 오류 고치는 중
ui에 기능을 넣는 중 오류가 생겨 고치는 중에 있습니다!
