+++ client/src/App.js
@@ -1,26 +1,13 @@
 import React from 'react';
-import styled from 'styled-components';
-import SearchBox from './components/SearchBox';
-import Recommendation from './components/Recommendation';
-import Search from './Search';
-
+import Search from './components/seachComponent/Search';
 
 function App() {
   return (
-    <Container>
-      <SearchBox />
-      <Recommendation />
-      <Search/>
-    </Container>
+    <>
+    ddd
+    <Search/>
+    </>
   );
 }
 export default App;
 
-const Container = styled.div`
-  background-color: #d0e8fd;
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-`;
